dashboard-admin/backend/auth/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
python-dotenv
python-multipart
passlib[bcrypt]